build(TLDExtractSwift.xcodeproj/project.pbxproj): Add alwaysOutOfDate flag to Update Public Suffix List script phase
This commit adds the 'alwaysOutOfDate' flag to the 'Update Public Suffix List' script phase in the TLDExtractSwift Xcode project file. This change ensures that the script will be executed every time the project is built, regardless of whether its output has changed or not. This is useful for scripts that perform tasks like updating data files, which should always be run to ensure the latest information is used.