build(TLDExtractSwift.xcodeproj/project.pbxproj): Add Public attribute to TLDExtractSwift.h in Headers and fix the error below.
`DEFINES_MODULE was set, but no umbrella header could be found to generate the module map`

The change adds the Public attribute to the TLDExtractSwift.h file located in the Headers folder of the TLDExtractSwift project. This attribute is used to specify that the header file can be accessed by other modules, making it public. The change was made to improve the accessibility and usability of the TLDExtractSwift library.